Move generic stream component from reader/following-stream to reader/stream
