+++ assets/stylesheets/_components.scss
@@ -345,7 +345,7 @@
 @import 'reader/discover/style';
 @import 'reader/feed-header/style';
 @import 'reader/following-edit/style';
-@import 'reader/following-stream/style';
+@import 'reader/stream/style';
 @import 'reader/full-post/style';
 @import 'reader/list-gap/style';
 @import 'reader/list-item/style';
