+++ client/reader/following/main.jsx
@@ -1,0 +1,13 @@
+// External dependencies
+import React from 'react';
+
+// Internal dependencies
+import Stream from 'reader/stream';
+
+const FollowingStream = ( props ) => {
+	return (
+		<Stream {...props} />
+	);
+};
+
+export default FollowingStream;
