+++ client/reader/liked-stream/main.jsx
@@ -1,6 +1,6 @@
 var React = require( 'react' );
 
-var FollowingStream = require( 'reader/following-stream' ),
+var Stream = require( 'reader/stream' ),
 	EmptyContent = require( './empty' );
 
 var LikedStream = React.createClass( {
@@ -13,7 +13,7 @@ var LikedStream = React.createClass( {
 			this.props.setPageTitle( title );
 		}
 		return (
-			<FollowingStream { ...this.props } listName={ title } emptyContent={ emptyContent } showFollowInHeader={ true } />
+			<Stream { ...this.props } listName={ title } emptyContent={ emptyContent } showFollowInHeader={ true } />
 		);
 	}
