+++ client/reader/stream/README.md
