+++ client/reader/stream/index.jsx
